added rabbitmq to vagrant. upgraded postgres to 9.5

you'll need to rebuild your vagrant box

See merge request !16

+++ vagrant/setup.sh
@@ -9,7 +9,21 @@ if [[ ! $(sudo -u postgres psql -t -c "SELECT datname FROM pg_database WHERE dat
     sudo -u postgres createdb serapis_dev -O serapis
 fi
 
-sudo apt-get -y install apache2
-sudo cp /vagrant/vagrant/conf/serapis-frontend.conf /etc/apache2/sites-enabled/
-sudo a2enmod headers
-sudo systemctl reload apache2
+if [ ! -e /sub/sbin/apache2 ]; then
+    sudo apt-get -y install apache2
+    sudo cp /vagrant/vagrant/conf/serapis-frontend.conf /etc/apache2/sites-enabled/
+    sudo a2enmod headers
+    sudo systemctl reload apache2
+fi
+
+install rabbitmq
+
+sudo rabbitmqctl list_vhosts | grep "^/serapis$" -q || {
+    sudo rabbitmqctl add_vhost /serapis
+    #give admin use access to this vhost in the managment interface
+    sudo rabbitmqctl set_permissions -p /serapis admin ".*" ".*" ".*"
+}
+sudo rabbitmqctl list_users | grep "^inserter\b" -q || {
+    sudo rabbitmqctl add_user inserter reallysecure
+    sudo rabbitmqctl set_permissions -p /serapis inserter ".*" ".*" ".*"
+}
